requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
defusedxml==0.7.1
Django==5.0.7
django-celery-beat==2.6.0
django-environ==0.11.2
django-rest-framework==0.1.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
httpie==3.2.3
idna==3.7
iniconfig==2.0.0
kombu==5.3.7
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
packaging==24.1
pluggy==1.5.0
podman-compose==1.2.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
Pygments==2.18.0
PyJWT==2.8.0
PySocks==1.7.1
pytest==8.3.1
pytest-django==4.8.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
redis==5.0.7
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
setuptools==70.3.0
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
